# Optimize performance for large numbers of services