--- a/pilot/pkg/model/push_context.go
+++ b/pilot/pkg/model/push_context.go
@@ -83,6 +83,7 @@ type PushContext struct {
 
 	// ServiceByHostnameAndNamespace has all services, indexed by hostname then namespace.
 	ServiceByHostnameAndNamespace map[host.Name]map[string]*Service `json:"-"`
+	ServiceByHostname             map[host.Name]*Service            `json:"-"`
 	// ServiceAccounts contains a map of hostname and port to service accounts.
 	ServiceAccounts map[host.Name]map[int][]string `json:"-"`
 	// QuotaSpec has all quota specs
@@ -456,6 +457,7 @@ func NewPushContext() *PushContext {
 		gatewaysByNamespace:                         map[string][]Config{},
 		allGateways:                                 []Config{},
 		ServiceByHostnameAndNamespace:               map[host.Name]map[string]*Service{},
+		ServiceByHostname:                           map[host.Name]*Service{},
 		ProxyStatus:                                 map[string]map[string]ProxyPushStatus{},
 		ServiceAccounts:                             map[host.Name]map[int][]string{},
 	}
@@ -899,6 +901,7 @@ func (ps *PushContext) updateContext(
 		ps.servicesExportedToNamespace = oldPushContext.servicesExportedToNamespace
 		ps.publicServices = oldPushContext.publicServices
 		ps.ServiceByHostnameAndNamespace = oldPushContext.ServiceByHostnameAndNamespace
+		ps.ServiceByHostname = oldPushContext.ServiceByHostname
 		ps.ServiceAccounts = oldPushContext.ServiceAccounts
 	}
 
@@ -1023,6 +1026,7 @@ func (ps *PushContext) initServiceRegistry(env *Environment) error {
 			ps.ServiceByHostnameAndNamespace[s.Hostname] = map[string]*Service{}
 		}
 		ps.ServiceByHostnameAndNamespace[s.Hostname][s.Attributes.Namespace] = s
+		ps.ServiceByHostname[s.Hostname] = s
 	}
 
 	ps.initServiceAccounts(env, allServices)