@istio-testing 2020-07-08
@howardjohn: The following test **failed**, say `/retest` to rerun all failed tests:

Test name | Commit | Details | Rerun command
--- | --- | --- | ---
integ-ipv6-k8s-tests_istio | 2749ec1f6bcec68c35491b8c0bdee92074693990 | [link](https://prow.istio.io/view/gcs/istio-prow/pr-logs/pull/istio_istio/25294/integ-ipv6-k8s-tests_istio/1732) | `/test integ-ipv6-k8s-tests_istio`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes/test-infra](https://github.com/kubernetes/test-infra/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->